requirements.txt
discord-py-interactions~=5.0.1
configparser~=5.3.0
mcrcon~=0.7.0
requests~=2.28.2
ping3~=4.0.4
PyYAML~=6.0
paramiko~=3.1.0
schedule~=1.2.0